// frontend/src/api/models/ChatbotTemplateOutputDTO.ts
export interface ChatbotTemplateOutputDTO {
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotName?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotModel?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotIcon?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotLanguage?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotRole?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    chatbotTone?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    welcomeMessage?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatbotTemplateOutputDTO
     */
    workspaceId?: string;
    /**
     * 
     * @type {Date}
     * @memberof ChatbotTemplateOutputDTO
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ChatbotTemplateOutputDTO
     */
    updatedAt?: Date;
}